Fix Shutdown OAuth Register
With the latest changes in 0.1.12 the `register_shutdown_handler` was always registered. This caused a problem in some scenarios.

When using the IBind client inside a Django application, it's not possible (or at least not trivial) to register a `SIGTERM` handler, since those are only supported in the main thread.

The change in this MR simply recovers the support to use `oauth_config.shutdown_oauth` value to both, register and execute the shutdown handler.